server/pyproject.toml
[project]
name = "edugenius-server"
version = "0.1.0"
description = "Backend for EduGenius AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlalchemy",
    "alembic",
    "asyncpg",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "httpx",
    "firebase-admin",
    "bcrypt",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]